Add endpoints for updating payment status and repo service for it

+++ src/main/java/dev/salt/Ring20/service/OrganizationApplicationService.java
@@ -26,6 +26,11 @@ public class OrganizationApplicationService {
     @Transactional
     public OrganizationApplication createApplication(String clerkId, String orgName, String description, String motivation) {
         OrganizationApplication application = new OrganizationApplication();
+        try {
+            Long userId = Long.valueOf(clerkId);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("Invalid user ID");
+        }
         User user = userService.getUserById(Long.valueOf(clerkId));
         application.setUser(user);
         application.setOrganizationName(orgName);
@@ -42,24 +47,42 @@ public class OrganizationApplicationService {
     }
 
     public OrganizationApplication getById(Long id) {
-        return repo.findById(id)
-                .orElseThrow(()-> new NoSuchElementException("No application found with id: "+ id));
-    }
-
-    @Transactional
-    public void approve(Long id) {
-        getById(id).
-                setApplicationStatus(ApplicationStatus.APPROVED);
-    }
-
-    @Transactional
-    public void reject(Long id) {
-        getById(id).
-                setApplicationStatus(ApplicationStatus.REJECTED);
+        return repo.findById(id).orElseThrow(() -> new NoSuchElementException("No application found with id: " + id));
     }
 
     @Transactional
     public void delete(Long id) {
+        if (!repo.existsById(id)) {
+            throw new NoSuchElementException("No application found with id: " + id);
+        }
         repo.deleteById(id);
+    }
+
+    @Transactional
+    public OrganizationApplication approve(Long id) {
+        OrganizationApplication application = getById(id);
+        application.setApplicationStatus(ApplicationStatus.APPROVED);
+        setReviewedTime(application);
+        return repo.save(application);
+    }
+
+    @Transactional
+    public OrganizationApplication reject(Long id) {
+        OrganizationApplication application = getById(id);
+        application.setApplicationStatus(ApplicationStatus.REJECTED);
+        setReviewedTime(application);
+        return repo.save(application);
+    }
+
+
+    private void setReviewedTime(OrganizationApplication application) {
+        application.setReviewedAt(LocalDateTime.now());
+    }
+
+    public OrganizationApplication updatePaymentStatus(Long id, PaymentStatus status) {
+        OrganizationApplication application = getById(id);
+        application.setPaymentStatus(status);
+        return repo.save(application);
+
     }
 }
